Fixed the linking of promote this dapp
